+++ Pwn/codebreaker/README.md
@@ -1,0 +1,10 @@
+# Codebreaker
+> Pwn - 100pts
+
+After saving the crew you've been captured by the Xenithians, and your escape depends on predicting the next three numbers in their code. Luckily, the Xenithians seem to have left out a crucial vulnerability when generating their code. Crack the code to be set free!
+
+Files:
+- [codebreaker.c](src/codebreaker.c)
+- [codebreaker](src/codebreaker)
+
+[Writeup](writeup/README.md)
