+++ Pwn/codebreaker/src/Makefile
@@ -1,0 +1,5 @@
+codebreaker:
+	gcc codebreaker.c -o codebreaker
+
+clean:
+	rm -f codebreaker
