--- Pwn/prng/src/Makefile
@@ -1,5 +1,0 @@
-prng:
-	gcc prng.c -o prng
-
-clean:
-	rm prng
