--- Pwn/prng/src/flag.txt
@@ -1,1 +1,0 @@
-UiTHack24{t3st_fl4g}
